migrate to config from the shared lib
