+++ .devcontainer/.sandbox_storage.yaml
@@ -1,6 +1,8 @@
 host: "127.0.0.1"
 port: 8080
-log_level: "debug"
+log_level: info
+drs_path: drs://localhost:8080/
+api_path: /ga4gh/drs/v1
 cors_allowed_origins: ["*"]
 cors_allow_credentials: true
 cors_allowed_methods: ["*"]
